Moving name to main ExprTranslator
